abstract out safetyWrapCoffeescript
