--- npm-debug.log
@@ -1,45 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ '/usr/local/bin/node', '/usr/local/bin/npm', 'run', 'lint' ]
-2 info using npm@3.10.9
-3 info using node@v7.1.0
-4 verbose run-script [ 'prelint', 'lint', 'postlint' ]
-5 info lifecycle drift@0.1.0~prelint: drift@0.1.0
-6 silly lifecycle drift@0.1.0~prelint: no script for prelint, continuing
-7 info lifecycle drift@0.1.0~lint: drift@0.1.0
-8 verbose lifecycle drift@0.1.0~lint: unsafe-perm in lifecycle true
-9 verbose lifecycle drift@0.1.0~lint: PATH: /usr/local/lib/node_modules/npm/bin/node-gyp-bin:/Users/m/workspace/drift/node_modules/.bin:/Users/m/.rvm/gems/ruby-2.3.1/bin:/Users/m/.rvm/gems/ruby-2.3.1@global/bin:/Users/m/.rvm/rubies/ruby-2.3.1/bin:/Users/m/miniconda2/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/opt/X11/bin:/usr/local/go/bin:/Users/m/.rvm/bin
-10 verbose lifecycle drift@0.1.0~lint: CWD: /Users/m/workspace/drift
-11 silly lifecycle drift@0.1.0~lint: Args: [ '-c', 'eslint index.js src --fix' ]
-12 silly lifecycle drift@0.1.0~lint: Returned: code: 1  signal: null
-13 info lifecycle drift@0.1.0~lint: Failed to exec lint script
-14 verbose stack Error: drift@0.1.0 lint: `eslint index.js src --fix`
-14 verbose stack Exit status 1
-14 verbose stack     at EventEmitter.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/lifecycle.js:255:16)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at EventEmitter.emit (events.js:191:7)
-14 verbose stack     at ChildProcess.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/spawn.js:40:14)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at ChildProcess.emit (events.js:191:7)
-14 verbose stack     at maybeClose (internal/child_process.js:885:16)
-14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
-15 verbose pkgid drift@0.1.0
-16 verbose cwd /Users/m/workspace/drift
-17 error Darwin 15.6.0
-18 error argv "/usr/local/bin/node" "/usr/local/bin/npm" "run" "lint"
-19 error node v7.1.0
-20 error npm  v3.10.9
-21 error code ELIFECYCLE
-22 error drift@0.1.0 lint: `eslint index.js src --fix`
-22 error Exit status 1
-23 error Failed at the drift@0.1.0 lint script 'eslint index.js src --fix'.
-23 error Make sure you have the latest version of node.js and npm installed.
-23 error If you do, this is most likely a problem with the drift package,
-23 error not with npm itself.
-23 error Tell the author that this fails on your system:
-23 error     eslint index.js src --fix
-23 error You can get information on how to open an issue for this project with:
-23 error     npm bugs drift
-23 error Or if that isn't available, you can get their info via:
-23 error     npm owner ls drift
-23 error There is likely additional logging output above.
-24 verbose exit [ 1, true ]
