+++ src/flowCoffeescriptKernel/flowCoffeescriptKernel.js
@@ -1,23 +1,11 @@
+import { safetyWrapCoffeescript } from './safetyWrapCoffeescript';
+
 export function flowCoffeescriptKernel() {
   const lodash = window._;
   const Flow = window.Flow;
   const escodegen = window.escodegen;
   const esprima = window.esprima;
   const CoffeeScript = window.CoffeeScript;
-  const safetyWrapCoffeescript = guid => (cs, go) => {
-    const lines = cs
-      .replace(/[\n\r]/g, '\n') // normalize CR/LF
-      .split('\n'); // split into lines
-
-    // indent once
-    const block = lodash.map(lines, line => `  ${line}`);
-
-    // enclose in execute-immediate closure
-    block.unshift(`_h2o_results_[\'${guid}\'].result do ->`);
-
-    // join and proceed
-    return go(null, block.join('\n'));
-  };
   const compileCoffeescript = (cs, go) => {
     let error;
     try {
